Add optional retry on closed connection failures
If the remote peer closes the connection, the ibrowse process will
return sel_conn_closed until it exits. While we handle the ibrowse
process exiting, in recent version of ibrowse that exit doesn't happen
until at least 500ms after the connection has been detected as closed
or closing [0]. Thus, we have a 500ms window after a remote peer closes
the conenction where we might erroneously attempt to reuse that
connection.

This change adds an opt-in single retry in the case of a
sel_conn_closed error.

[0] https://github.com/chef/ibrowse/blob/7b2b0bfd5e15e57706beb2ef0360cdc7f24b37c5/src/ibrowse_http_client.erl#L2182-L2183

Signed-off-by: Steven Danna <steve@chef.io>