Add custom verify_ca TLS option
Erlang 19+ not verifies hostnames as part of it's TLS
implementation. This is generally what we want, but unfortunately by
default that verification does not work well with wildcard certs or
with certain kinds of subject alternative names.

The situation is better in Erlang 22, however, we are still working
through some issues upgrading our ecosystem to this version.

This change adds a temporary alternative verification mode `verify_ca`
which ignores hostname mismatches. We implemented a similar feature in
our SQL library since it aligns with PostgreSQLs verify_ca verify
mode.

Note we will likely remove this feature when we are able to upgrade to
Erlang 22.

Signed-off-by: Steven Danna <steve@chef.io>